Move group and menu to separate directory
